+++ src/main/java/org/spin/grpc/service/form/ExpressReceipt.java
@@ -1,5 +1,5 @@
 /************************************************************************************
- * Copyright (C) 2018-2023 E.R.P. Consultores y Asociados, C.A.                     *
+ * Copyright (C) 2018-present E.R.P. Consultores y Asociados, C.A.                  *
  * Contributor(s): Edwin Betancourt, EdwinBetanc0urt@outlook.com                    *
  * This program is free software: you can redistribute it and/or modify             *
  * it under the terms of the GNU General Public License as published by             *
@@ -12,7 +12,7 @@
  * You should have received a copy of the GNU General Public License                *
  * along with this program. If not, see <https://www.gnu.org/licenses/>.            *
  ************************************************************************************/
-package org.spin.grpc.service;
+package org.spin.grpc.service.form;
 
 import org.adempiere.exceptions.AdempiereException;
 
@@ -32,37 +32,39 @@ import org.adempiere.core.domains.models.I_M_InOutLine;
 import org.adempiere.core.domains.models.I_M_Product;
 import org.adempiere.core.domains.models.X_M_InOut;
 import org.compiere.model.MBPartner;
+import org.compiere.model.MDocType;
 import org.compiere.model.MInOut;
 import org.compiere.model.MInOutLine;
 import org.compiere.model.MOrder;
 import org.compiere.model.MOrderLine;
 import org.compiere.model.MProduct;
 import org.compiere.model.MRole;
+import org.compiere.model.MSysConfig;
 import org.compiere.model.Query;
 import org.compiere.util.CLogger;
 import org.compiere.util.Env;
 import org.compiere.util.Trx;
 import org.compiere.util.Util;
-import org.spin.backend.grpc.form.express_shipment.BusinessPartner;
-import org.spin.backend.grpc.form.express_shipment.CreateShipmentLineRequest;
-import org.spin.backend.grpc.form.express_shipment.CreateShipmentRequest;
-import org.spin.backend.grpc.form.express_shipment.DeleteShipmentLineRequest;
-import org.spin.backend.grpc.form.express_shipment.DeleteShipmentRequest;
-import org.spin.backend.grpc.form.express_shipment.ListBusinessPartnersRequest;
-import org.spin.backend.grpc.form.express_shipment.ListBusinessPartnersResponse;
-import org.spin.backend.grpc.form.express_shipment.ListProductsRequest;
-import org.spin.backend.grpc.form.express_shipment.ListProductsResponse;
-import org.spin.backend.grpc.form.express_shipment.ListSalesOrdersRequest;
-import org.spin.backend.grpc.form.express_shipment.ListSalesOrdersResponse;
-import org.spin.backend.grpc.form.express_shipment.ListShipmentLinesRequest;
-import org.spin.backend.grpc.form.express_shipment.ListShipmentLinesResponse;
-import org.spin.backend.grpc.form.express_shipment.ProcessShipmentRequest;
-import org.spin.backend.grpc.form.express_shipment.Product;
-import org.spin.backend.grpc.form.express_shipment.SalesOrder;
-import org.spin.backend.grpc.form.express_shipment.Shipment;
-import org.spin.backend.grpc.form.express_shipment.ShipmentLine;
-import org.spin.backend.grpc.form.express_shipment.UpdateShipmentLineRequest;
-import org.spin.backend.grpc.form.express_shipment.ExpressShipmentGrpc.ExpressShipmentImplBase;
+import org.spin.backend.grpc.form.express_receipt.BusinessPartner;
+import org.spin.backend.grpc.form.express_receipt.CreateReceiptLineRequest;
+import org.spin.backend.grpc.form.express_receipt.CreateReceiptRequest;
+import org.spin.backend.grpc.form.express_receipt.DeleteReceiptLineRequest;
+import org.spin.backend.grpc.form.express_receipt.DeleteReceiptRequest;
+import org.spin.backend.grpc.form.express_receipt.ListBusinessPartnersRequest;
+import org.spin.backend.grpc.form.express_receipt.ListBusinessPartnersResponse;
+import org.spin.backend.grpc.form.express_receipt.ListProductsRequest;
+import org.spin.backend.grpc.form.express_receipt.ListProductsResponse;
+import org.spin.backend.grpc.form.express_receipt.ListPurchaseOrdersRequest;
+import org.spin.backend.grpc.form.express_receipt.ListPurchaseOrdersResponse;
+import org.spin.backend.grpc.form.express_receipt.ListReceiptLinesRequest;
+import org.spin.backend.grpc.form.express_receipt.ListReceiptLinesResponse;
+import org.spin.backend.grpc.form.express_receipt.ProcessReceiptRequest;
+import org.spin.backend.grpc.form.express_receipt.Product;
+import org.spin.backend.grpc.form.express_receipt.PurchaseOrder;
+import org.spin.backend.grpc.form.express_receipt.Receipt;
+import org.spin.backend.grpc.form.express_receipt.ReceiptLine;
+import org.spin.backend.grpc.form.express_receipt.UpdateReceiptLineRequest;
+import org.spin.backend.grpc.form.express_receipt.ExpressReceiptGrpc.ExpressReceiptImplBase;
 import org.spin.base.db.LimitUtil;
 import org.spin.base.util.DocumentUtil;
 import org.spin.base.util.RecordUtil;
@@ -77,11 +79,11 @@ import io.grpc.stub.StreamObserver;
 
 /**
  * @author Edwin Betancourt, EdwinBetanc0urt@outlook.com, https://github.com/EdwinBetanc0urt
- * Service for backend of Express Shipment
+ * Service for backend of Express Receipt
  */
-public class ExpressShipment extends ExpressShipmentImplBase {
+public class ExpressReceipt extends ExpressReceiptImplBase {
 	/**	Logger			*/
-	private CLogger log = CLogger.getCLogger(ExpressShipment.class);
+	private CLogger log = CLogger.getCLogger(ExpressReceipt.class);
 	
 	public String tableName = I_M_InOut.Table_Name;
 
@@ -98,6 +100,7 @@ public class ExpressShipment extends ExpressShipmentImplBase {
 			responseObserver.onCompleted();
 		} catch (Exception e) {
 			log.severe(e.getLocalizedMessage());
+			e.printStackTrace();
 			responseObserver.onError(Status.INTERNAL
 				.withDescription(e.getLocalizedMessage())
 				.withCause(e)
@@ -107,15 +110,15 @@ public class ExpressShipment extends ExpressShipmentImplBase {
 	}
 
 	ListBusinessPartnersResponse.Builder listBusinessPartners(ListBusinessPartnersRequest request) {
-		final String whereClause = "IsCustomer='Y' ";
+		final String whereClause = "IsVendor='Y' ";
 		Query query = new Query(
 			Env.getCtx(),
 			I_C_BPartner.Table_Name,
 			whereClause,
 			null
 		)
-			.setClient_ID()
 			.setOnlyActiveRecords(true)
+			.setClient_ID()
 			.setApplyAccessFilter(MRole.SQL_FULLYQUALIFIED, MRole.SQL_RO)
 		;
 
@@ -135,8 +138,9 @@ public class ExpressShipment extends ExpressShipmentImplBase {
 		;
 
 		query.setLimit(limit, offset)
-			.list(MBPartner.class)
-			.forEach(businessPartner -> {
+			.getIDsAsList()
+			.forEach(businessPartnerId -> {
+				MBPartner businessPartner = MBPartner.get(Env.getCtx(), businessPartnerId);
 				BusinessPartner.Builder builder = convertBusinessPartner(businessPartner);
 				builderList.addRecords(builder);
 			})
@@ -147,7 +151,7 @@ public class ExpressShipment extends ExpressShipmentImplBase {
 
 	BusinessPartner.Builder convertBusinessPartner(MBPartner businessPartner) {
 		BusinessPartner.Builder builder = BusinessPartner.newBuilder();
-		if (builder == null) {
+		if (businessPartner == null) {
 			return builder;
 		}
 
@@ -171,17 +175,18 @@ public class ExpressShipment extends ExpressShipmentImplBase {
 
 
 	@Override
-	public void listSalesOrders(ListSalesOrdersRequest request, StreamObserver<ListSalesOrdersResponse> responseObserver) {
+	public void listPurchaseOrders(ListPurchaseOrdersRequest request, StreamObserver<ListPurchaseOrdersResponse> responseObserver) {
 		try {
 			if (request == null) {
 				throw new AdempiereException("Object Request Null");
 			}
 
-			ListSalesOrdersResponse.Builder builderList = listSalesOrders(request);
+			ListPurchaseOrdersResponse.Builder builderList = listPurchaseOrders(request);
 			responseObserver.onNext(builderList.build());
 			responseObserver.onCompleted();
 		} catch (Exception e) {
 			log.severe(e.getLocalizedMessage());
+			e.printStackTrace();
 			responseObserver.onError(Status.INTERNAL
 				.withDescription(e.getLocalizedMessage())
 				.withCause(e)
@@ -195,20 +200,21 @@ public class ExpressShipment extends ExpressShipmentImplBase {
 	 * @param request
 	 * @return
 	 */
-	private ListSalesOrdersResponse.Builder listSalesOrders(ListSalesOrdersRequest request) {
+	private ListPurchaseOrdersResponse.Builder listPurchaseOrders(ListPurchaseOrdersRequest request) {
 		Properties context = Env.getCtx();
 
 		int businessPartnerId = request.getBusinessPartnerId();
 		if (businessPartnerId <= 0) {
-			throw new AdempiereException("@C_BPartner_ID@ @NotFound@");
+			throw new AdempiereException("@FillMandatory@ @C_BPartner_ID@");
 		}
 
-		final String whereClause = "IsSOTrx='Y' "
+		// See dynamic validation 52464
+		final String whereClause = "IsSOTrx='N' "
 			+ "AND C_BPartner_ID = ? "
-			+ "AND DocStatus IN ('CL','CO')"
+			+ "AND DocStatus IN ('CL','CO') "
 			+ "AND EXISTS(SELECT 1 FROM C_OrderLine ol "
 			+ "WHERE ol.C_Order_ID = C_Order.C_Order_ID "
-			+ "AND COALESCE(ol.QtyOrdered, 0) - COALESCE(ol.QtyDelivered, 0) > 0)"
+			+ "AND ol.QtyOrdered - ol.QtyDelivered > 0)"
 		;
 		Query query = new Query(
 			context,
@@ -231,34 +237,37 @@ public class ExpressShipment extends ExpressShipmentImplBase {
 			nexPageToken = LimitUtil.getPagePrefix(SessionManager.getSessionUuid()) + (pageNumber + 1);
 		}
 
-		ListSalesOrdersResponse.Builder builderList = ListSalesOrdersResponse.newBuilder()
+		ListPurchaseOrdersResponse.Builder builderList = ListPurchaseOrdersResponse.newBuilder()
 			.setRecordCount(count)
 			.setNextPageToken(nexPageToken)
 		;
 
 		query.setLimit(limit, offset)
-			.list(MOrder.class)
-			.forEach(salesOrder -> {
-				SalesOrder.Builder builder = convertSalesOrder(salesOrder);
+			.getIDsAsList()
+			.forEach(purchaseOrderId -> {
+				MOrder purcharseOrder = new MOrder(context, purchaseOrderId, null);
+				PurchaseOrder.Builder builder = convertPurchaseOrder(
+					purcharseOrder
+				);
 				builderList.addRecords(builder);
-			});
+			})
 		;
 
 		return builderList;
 	}
 
-	SalesOrder.Builder convertSalesOrder(MOrder salesOrder) {
-		SalesOrder.Builder builder = SalesOrder.newBuilder();
-		if (salesOrder == null) {
+	PurchaseOrder.Builder convertPurchaseOrder(MOrder purchaseOrder) {
+		PurchaseOrder.Builder builder = PurchaseOrder.newBuilder();
+		if (purchaseOrder == null) {
 			return builder;
 		}
 
-		builder.setId(salesOrder.getC_Order_ID())
+		builder.setId(purchaseOrder.getC_Order_ID())
 			.setDateOrdered(
-				ValueManager.getTimestampFromDate(salesOrder.getDateOrdered())
+				ValueManager.getTimestampFromDate(purchaseOrder.getDateOrdered())
 			)
 			.setDocumentNo(
-				ValueManager.validateNull(salesOrder.getDocumentNo())
+				ValueManager.validateNull(purchaseOrder.getDocumentNo())
 			)
 		;
 
@@ -290,11 +299,11 @@ public class ExpressShipment extends ExpressShipmentImplBase {
 	private ListProductsResponse.Builder listProducts(ListProductsRequest request) {
 		int orderId = request.getOrderId();
 		if (orderId <= 0) {
-			throw new AdempiereException("@C_Order_ID@ @NotFound@");
+			throw new AdempiereException("@FillMandatory@ @C_Order_ID@");
 		}
 
 		//	Dynamic where clause
-		String whereClause = "IsSold = 'Y' AND IsSummary = 'N' "
+		String whereClause = "IsPurchased = 'Y' AND IsSummary = 'N' "
 			+ "AND EXISTS(SELECT 1 FROM C_OrderLine AS OL "
 			+ "WHERE OL.C_Order_ID = ? AND OL.M_Product_ID = M_Product.M_Product_ID) "
 		;
@@ -347,9 +356,9 @@ public class ExpressShipment extends ExpressShipmentImplBase {
 		;
 
 		query.setLimit(limit, offset)
-			.list(MProduct.class)
-			.forEach(product -> {
-				Product.Builder builder = convertProduct(product);
+			.getIDsAsList()
+			.forEach(productId -> {
+				Product.Builder builder = convertProduct(productId);
 				builderList.addRecords(builder);
 			});
 		;
@@ -389,29 +398,29 @@ public class ExpressShipment extends ExpressShipmentImplBase {
 	}
 
 
-	Shipment.Builder convertShipment(MInOut shipment) {
-		Shipment.Builder builder = Shipment.newBuilder();
-		if (shipment == null) {
+	Receipt.Builder convertReceipt(MInOut receipt) {
+		Receipt.Builder builder = Receipt.newBuilder();
+		if (receipt == null) {
 			return builder;
 		}
 
-		builder.setId(shipment.getM_InOut_ID())
+		builder.setId(receipt.getM_InOut_ID())
 			.setDocumentNo(
-				ValueManager.validateNull(shipment.getDocumentNo())
+				ValueManager.validateNull(receipt.getDocumentNo())
 			)
 			.setDateOrdered(
-				ValueManager.getTimestampFromDate(shipment.getDateOrdered())
+				ValueManager.getTimestampFromDate(receipt.getDateOrdered())
 			)
 			.setMovementDate(
-				ValueManager.getTimestampFromDate(shipment.getMovementDate())
+				ValueManager.getTimestampFromDate(receipt.getMovementDate())
 			)
 			.setIsCompleted(
-				DocumentUtil.isCompleted(shipment)
+				DocumentUtil.isCompleted(receipt)
 			)
 		;
-		MOrderLine salesOrderLine = new MOrderLine(Env.getCtx(), shipment.getC_Order_ID(), null);
-		if (salesOrderLine != null && salesOrderLine.getC_OrderLine_ID() > 0) {
-			builder.setOrderId(salesOrderLine.getC_OrderLine_ID())
+		MOrderLine purchaseOrderLine = new MOrderLine(Env.getCtx(), receipt.getC_Order_ID(), null);
+		if (purchaseOrderLine != null && purchaseOrderLine.getC_OrderLine_ID() > 0) {
+			builder.setOrderId(purchaseOrderLine.getC_OrderLine_ID())
 			;
 		}
 
@@ -420,17 +429,18 @@ public class ExpressShipment extends ExpressShipmentImplBase {
 
 
 	@Override
-	public void createShipment(CreateShipmentRequest request, StreamObserver<Shipment> responseObserver) {
+	public void createReceipt(CreateReceiptRequest request, StreamObserver<Receipt> responseObserver) {
 		try {
 			if (request == null) {
 				throw new AdempiereException("Object Request Null");
 			}
 
-			Shipment.Builder builder = createShipment(request);
+			Receipt.Builder builder = createReceipt(request);
 			responseObserver.onNext(builder.build());
 			responseObserver.onCompleted();
 		} catch (Exception e) {
 			log.severe(e.getLocalizedMessage());
+			e.printStackTrace();
 			responseObserver.onError(Status.INTERNAL
 				.withDescription(e.getLocalizedMessage())
 				.withCause(e)
@@ -439,99 +449,114 @@ public class ExpressShipment extends ExpressShipmentImplBase {
 		}
 	}
 
-	private Shipment.Builder createShipment(CreateShipmentRequest request) {
-		if (request.getOrderId() <= 0) {
-			throw new AdempiereException("@C_Order_ID@ @NotFound@");
+	private Receipt.Builder createReceipt(CreateReceiptRequest request) {
+		int orderId = request.getOrderId();
+		if (orderId <= 0) {
+			throw new AdempiereException("@FillMandatory@ @C_Order_ID@");
 		}
 
-		AtomicReference<MInOut> maybeShipment = new AtomicReference<MInOut>();
+		AtomicReference<MInOut> maybeReceipt = new AtomicReference<MInOut>();
 		Trx.run(transactionName -> {
-			int orderId = request.getOrderId();
-			if (orderId <= 0) {
+			MOrder purchaseOrder = new MOrder(Env.getCtx(), orderId, transactionName);
+			if (purchaseOrder == null || purchaseOrder.getC_Order_ID() <= 0) {
 				throw new AdempiereException("@C_Order_ID@ @NotFound@");
 			}
-			MOrder salesOrder = new MOrder(Env.getCtx(), orderId, transactionName);
-			if (salesOrder == null || salesOrder.getC_Order_ID() <= 0) {
-				throw new AdempiereException("@C_Order_ID@ @NotFound@");
-			}
-			// if (salesOrder.isDelivered()) {
+			// if (purchaseOrder.isDelivered()) {
 			// 	throw new AdempiereException("@C_Order_ID@ @IsDelivered@");
 			// }
 			// Valid if has a Order
-			if (!DocumentUtil.isCompleted(salesOrder)) {
-				throw new AdempiereException("@Invalid@ @C_Order_ID@ " + salesOrder.getDocumentNo());
+			if (!DocumentUtil.isCompleted(purchaseOrder)) {
+				throw new AdempiereException("@Invalid@ @C_Order_ID@ " + purchaseOrder.getDocumentNo());
+			}
+			MDocType orderDocType = MDocType.get(Env.getCtx(), purchaseOrder.getC_DocType_ID());
+			if (orderDocType.getC_DocTypeShipment_ID() <= 0) {
+				log.warning(
+					"@C_Order_ID@ " + purchaseOrder.getDocumentNo()
+					+ " @C_DocType_ID@ " + orderDocType.getName()
+					+ " : @C_DocTypeShipment_ID@ @NotFound@"
+				);
 			}
 
-			final String whereClause = "IsSOTrx='Y' "
-				+ "AND MovementType IN ('V-') "
+			final String whereClause = "IsSOTrx='N' "
+				+ "AND MovementType IN ('V+') "
 				+ "AND DocStatus = 'DR' "
 				+ "AND C_Order_ID = ? "
 			;
-
-			MInOut shipment = new Query(
+			MInOut receipt = new Query(
 				Env.getCtx(),
 				I_M_InOut.Table_Name,
 				whereClause,
 				transactionName
 			)
-				.setParameters(salesOrder.getC_Order_ID())
+				.setParameters(purchaseOrder.getC_Order_ID())
 				.setClient_ID()
 				.first();
 
 			// Validate
-			if (shipment == null) {
-				shipment = new MInOut(salesOrder, 0, RecordUtil.getDate());
+			if (receipt == null) {
+				receipt = new MInOut(purchaseOrder, 0, RecordUtil.getDate());
+				if (Util.isEmpty(receipt.getMovementType())) {
+					receipt.setMovementType(X_M_InOut.MOVEMENTTYPE_VendorReceipts);
+				}
 			} else {
-				if(!DocumentUtil.isDrafted(shipment)) {
+				if(!DocumentUtil.isDrafted(receipt)) {
 					throw new AdempiereException("@M_InOut_ID@ @Processed@");
 				}
-				shipment.setDateOrdered(RecordUtil.getDate());
-				shipment.setDateAcct(RecordUtil.getDate());
-				shipment.setDateReceived(RecordUtil.getDate());
-				shipment.setMovementDate(RecordUtil.getDate());
+				receipt.setDateOrdered(RecordUtil.getDate());
+				receipt.setDateAcct(RecordUtil.getDate());
+				receipt.setDateReceived(RecordUtil.getDate());
+				receipt.setMovementDate(RecordUtil.getDate());
 			}
-			shipment.setC_Order_ID(orderId);
-			shipment.saveEx(transactionName);
+			receipt.setC_Order_ID(orderId);
+			receipt.saveEx(transactionName);
 
-			maybeShipment.set(shipment);
+			maybeReceipt.set(receipt);
 
 			if (request.getIsCreateLinesFromOrder()) {
-				List<MOrderLine> orderLines = Arrays.asList(salesOrder.getLines());
-				orderLines.stream().forEach(salesOrderLine -> {
-					Optional<MInOutLine> maybeShipmentLine = Arrays.asList(maybeShipment.get().getLines(true))
+				boolean validateOrderedQty = MSysConfig.getBooleanValue(
+					"VALIDATE_MATCHING_TO_ORDERED_QTY",
+					true,
+					Env.getAD_Client_ID(Env.getCtx())
+				);
+
+				List<MOrderLine> orderLines = Arrays.asList(purchaseOrder.getLines());
+				orderLines.stream().forEach(purchaseOrderLine -> {
+					Optional<MInOutLine> maybeReceiptLine = Arrays.asList(maybeReceipt.get().getLines(true))
 						.stream()
 						.filter(shipmentLineTofind -> {
-							return shipmentLineTofind.getC_OrderLine_ID() == salesOrderLine.getC_OrderLine_ID();
+							return shipmentLineTofind.getC_OrderLine_ID() == purchaseOrderLine.getC_OrderLine_ID();
 						})
 						.findFirst();
 
-					if (maybeShipmentLine.isPresent()) {
-						MInOutLine shipmentLine = maybeShipmentLine.get();
+					if (maybeReceiptLine.isPresent()) {
+						MInOutLine receiptLine = maybeReceiptLine.get();
 
-						BigDecimal quantity = salesOrderLine.getQtyEntered();
+						BigDecimal quantity = purchaseOrderLine.getQtyEntered();
 						if (quantity == null) {
-							quantity = shipmentLine.getMovementQty().add(Env.ONE);
+							quantity = receiptLine.getMovementQty().add(Env.ONE);
 						}
 
 						// Validate available
-						BigDecimal orderQuantityDelivered = salesOrderLine.getQtyOrdered().subtract(salesOrderLine.getQtyDelivered());
-						if (orderQuantityDelivered.compareTo(quantity) < 0) {
-							throw new AdempiereException("@QtyInsufficient@");
+						if (validateOrderedQty) {
+							BigDecimal orderQuantityDelivered = purchaseOrderLine.getQtyOrdered().subtract(purchaseOrderLine.getQtyDelivered());
+							if (orderQuantityDelivered.compareTo(quantity) < 0) {
+								throw new AdempiereException("@QtyInsufficient@");
+							}
 						}
-						shipmentLine.setQty(quantity);
-						shipmentLine.saveEx();
+						receiptLine.setQty(quantity);
+						receiptLine.saveEx();
 					} else {
-						MInOutLine shipmentLine = new MInOutLine(maybeShipment.get());
+						MInOutLine receiptLine = new MInOutLine(maybeReceipt.get());
 
-						BigDecimal quantity = salesOrderLine.getQtyReserved();
-						shipmentLine.setOrderLine(salesOrderLine, 0, quantity);
-						shipmentLine.saveEx(transactionName);
+						BigDecimal quantity = purchaseOrderLine.getQtyReserved();
+						receiptLine.setOrderLine(purchaseOrderLine, 0, quantity);
+						receiptLine.saveEx(transactionName);
 					}
 				});
 			}
 		});
 
-		Shipment.Builder builder = convertShipment(maybeShipment.get());
+		Receipt.Builder builder = convertReceipt(maybeReceipt.get());
 
 		return builder;
 	}
@@ -539,17 +564,18 @@ public class ExpressShipment extends ExpressShipmentImplBase {
 
 
 	@Override
-	public void deleteShipment(DeleteShipmentRequest request, StreamObserver<Empty> responseObserver) {
+	public void deleteReceipt(DeleteReceiptRequest request, StreamObserver<Empty> responseObserver) {
 		try {
 			if (request == null) {
 				throw new AdempiereException("Object Request Null");
 			}
 
-			Empty.Builder builder = deleteShipment(request);
+			Empty.Builder builder = deleteReceipt(request);
 			responseObserver.onNext(builder.build());
 			responseObserver.onCompleted();
 		} catch (Exception e) {
 			log.severe(e.getLocalizedMessage());
+			e.printStackTrace();
 			responseObserver.onError(Status.INTERNAL
 				.withDescription(e.getLocalizedMessage())
 				.withCause(e)
@@ -558,24 +584,21 @@ public class ExpressShipment extends ExpressShipmentImplBase {
 		}
 	}
 
-	private Empty.Builder deleteShipment(DeleteShipmentRequest request) {
-		if (request.getId() <= 0) {
-			throw new AdempiereException("@M_InOut_ID@ @NotFound@");
+	private Empty.Builder deleteReceipt(DeleteReceiptRequest request) {
+		int receiptId = request.getId();
+		if (receiptId <= 0) {
+			throw new AdempiereException("@FillMandatory@ @M_InOut_ID@");
 		}
-		Trx.run(transactionName -> {
-			int shipmentId = request.getId();
-			if (shipmentId <= 0) {
-				throw new AdempiereException("@M_InOut_ID@ @NotFound@");
-			}
 
-			MInOut shipment = new MInOut(Env.getCtx(), shipmentId, transactionName);
-			if (shipment == null || shipment.getM_InOut_ID() <= 0) {
+		Trx.run(transactionName -> {
+			MInOut receipt = new MInOut(Env.getCtx(), receiptId, transactionName);
+			if (receipt == null || receipt.getM_InOut_ID() <= 0) {
 				throw new AdempiereException("@M_InOut_ID@ @NotFound@");
 			}
-			if (!DocumentUtil.isDrafted(shipment)) {
-				throw new AdempiereException("@Invalid@ @M_InOut_ID@ " + shipment.getDocumentNo());
+			if (!DocumentUtil.isDrafted(receipt)) {
+				throw new AdempiereException("@Invalid@ @M_InOut_ID@ " + receipt.getDocumentNo());
 			}
-			shipment.deleteEx(true);
+			receipt.deleteEx(true);
 		});
 
 		return Empty.newBuilder();
@@ -584,17 +607,18 @@ public class ExpressShipment extends ExpressShipmentImplBase {
 
 
 	@Override
-	public void processShipment(ProcessShipmentRequest request, StreamObserver<Shipment> responseObserver) {
+	public void processReceipt(ProcessReceiptRequest request, StreamObserver<Receipt> responseObserver) {
 		try {
 			if (request == null) {
 				throw new AdempiereException("Object Request Null");
 			}
 
-			Shipment.Builder builder = processShipment(request);
+			Receipt.Builder builder = processReceipt(request);
 			responseObserver.onNext(builder.build());
 			responseObserver.onCompleted();
 		} catch (Exception e) {
 			log.severe(e.getLocalizedMessage());
+			e.printStackTrace();
 			responseObserver.onError(Status.INTERNAL
 				.withDescription(e.getLocalizedMessage())
 				.withCause(e)
@@ -603,31 +627,31 @@ public class ExpressShipment extends ExpressShipmentImplBase {
 		}
 	}
 
-	private Shipment.Builder processShipment(ProcessShipmentRequest request) {
-		AtomicReference<MInOut> shipmentReference = new AtomicReference<MInOut>();
+	private Receipt.Builder processReceipt(ProcessReceiptRequest request) {
+		int receiptId = request.getId();
+		if (receiptId <= 0) {
+			throw new AdempiereException("@FillMandatory@ @M_InOut_ID@");
+		}
+
+		AtomicReference<MInOut> receiptReference = new AtomicReference<MInOut>();
 
 		Trx.run(transactionName -> {
-			int shipmentId = request.getId();
-			if (shipmentId <= 0) {
+			MInOut receipt = new MInOut(Env.getCtx(), receiptId, transactionName);
+			if (receipt == null || receiptId <= 0) {
 				throw new AdempiereException("@M_InOut_ID@ @NotFound@");
 			}
-
-			MInOut shipment = new MInOut(Env.getCtx(), shipmentId, transactionName);
-			if (shipment == null || shipmentId <= 0) {
-				throw new AdempiereException("@M_InOut_ID@ @NotFound@");
-			}
-			if (shipment.isProcessed()) {
+			if (receipt.isProcessed()) {
 				throw new AdempiereException("@M_InOut_ID@ @Processed@");
 			}
-			if (!shipment.processIt(X_M_InOut.DOCACTION_Complete)) {
-				log.warning("@ProcessFailed@ :" + shipment.getProcessMsg());
-				throw new AdempiereException("@ProcessFailed@ :" + shipment.getProcessMsg());
+			if (!receipt.processIt(X_M_InOut.DOCACTION_Complete)) {
+				log.warning("@ProcessFailed@ :" + receipt.getProcessMsg());
+				throw new AdempiereException("@ProcessFailed@ :" + receipt.getProcessMsg());
 			}
-			shipment.saveEx(transactionName);
-			shipmentReference.set(shipment);
+			receipt.saveEx(transactionName);
+			receiptReference.set(receipt);
 		});
 
-		Shipment.Builder builder = convertShipment(shipmentReference.get());
+		Receipt.Builder builder = convertReceipt(receiptReference.get());
 
 		return builder;
 	}
@@ -635,17 +659,18 @@ public class ExpressShipment extends ExpressShipmentImplBase {
 
 
 	@Override
-	public void createShipmentLine(CreateShipmentLineRequest request, StreamObserver<ShipmentLine> responseObserver) {
+	public void createReceiptLine(CreateReceiptLineRequest request, StreamObserver<ReceiptLine> responseObserver) {
 		try {
 			if (request == null) {
 				throw new AdempiereException("Object Request Null");
 			}
 
-			ShipmentLine.Builder builder = createShipmentLine(request);
+			ReceiptLine.Builder builder = createReceiptLine(request);
 			responseObserver.onNext(builder.build());
 			responseObserver.onCompleted();
 		} catch (Exception e) {
 			log.severe(e.getLocalizedMessage());
+			e.printStackTrace();
 			responseObserver.onError(Status.INTERNAL
 				.withDescription(e.getLocalizedMessage())
 				.withCause(e)
@@ -654,49 +679,43 @@ public class ExpressShipment extends ExpressShipmentImplBase {
 		}
 	}
 
-	private ShipmentLine.Builder createShipmentLine(CreateShipmentLineRequest request) {
-		if (request.getShipmentId() <= 0) {
-			throw new AdempiereException("@M_InOut_ID@ @NotFound@");
+	private ReceiptLine.Builder createReceiptLine(CreateReceiptLineRequest request) {
+		int receiptId = request.getReceiptId();
+		if (receiptId <= 0) {
+			throw new AdempiereException("@FillMandatory@ @M_InOut_ID@");
 		}
-		if (request.getProductId() <= 0) {
-			throw new AdempiereException("@M_Product_ID@ @NotFound@");
+		int productId = request.getProductId();
+		if (productId <= 0) {
+			throw new AdempiereException("@FillMandatory@ @M_Product_ID@");
 		}
 
-		AtomicReference<MInOutLine> shipmentLineReference = new AtomicReference<MInOutLine>();
+		AtomicReference<MInOutLine> receiptLineReference = new AtomicReference<MInOutLine>();
 
 		Trx.run(transactionName -> {
-			int shipmentId = request.getShipmentId();
-			if (shipmentId <= 0) {
+			MInOut receipt = new MInOut(Env.getCtx(), receiptId, transactionName);
+			if (receipt == null || receipt.getM_InOut_ID() <= 0) {
 				throw new AdempiereException("@M_InOut_ID@ @NotFound@");
 			}
-			MInOut shipment = new MInOut(Env.getCtx(), shipmentId, transactionName);
-			if (shipment == null || shipment.getM_InOut_ID() <= 0) {
-				throw new AdempiereException("@M_InOut_ID@ @NotFound@");
-			}
-			if (shipment.isProcessed()) {
+			if (receipt.isProcessed()) {
 				throw new AdempiereException("@M_InOut_ID@ @Processed@");
 			}
 
-			int productId = request.getProductId();
-			if (productId <= 0) {
-				throw new AdempiereException("@M_Product_ID@ @NotFound@");
-			}
 
 			final String whereClause = "M_Product_ID = ? "
 				+ "AND EXISTS(SELECT 1 FROM M_InOut "
 				+ "WHERE M_InOut.C_Order_ID = C_OrderLine.C_Order_ID "
 				+ "AND M_InOut.M_InOut_ID = ?)"
 			;
-			MOrderLine salesOrderLine = new Query(
+			MOrderLine purchaseOrderLine = new Query(
 				Env.getCtx(),
 				I_C_OrderLine.Table_Name,
 				whereClause,
 				transactionName
 			)
-				.setParameters(productId, shipmentId)
+				.setParameters(productId, receiptId)
 				.setClient_ID()
 				.first();
-			if (salesOrderLine == null || salesOrderLine.getC_OrderLine_ID() <= 0) {
+			if (purchaseOrderLine == null || purchaseOrderLine.getC_OrderLine_ID() <= 0) {
 				throw new AdempiereException("@C_OrderLine_ID@ @NotFound@");
 			}
 
@@ -710,40 +729,35 @@ public class ExpressShipment extends ExpressShipmentImplBase {
 				}
 			}
 			if (request.getIsQuantityFromOrderLine()) {
-				quantity = salesOrderLine.getQtyReserved();
-			}
-			// Validate available
-			BigDecimal orderQuantityDelivered = salesOrderLine.getQtyOrdered().subtract(salesOrderLine.getQtyDelivered());
-			if (orderQuantityDelivered.compareTo(quantity) < 0) {
-				throw new AdempiereException("@QtyInsufficient@");
+				quantity = purchaseOrderLine.getQtyReserved();
 			}
 
-			Optional<MInOutLine> maybeShipmentLine = Arrays.asList(shipment.getLines(true))
+			Optional<MInOutLine> maybeReceiptLine = Arrays.asList(receipt.getLines(true))
 				.stream()
-				.filter(shipmentLineTofind -> {
-					return shipmentLineTofind.getC_OrderLine_ID() == salesOrderLine.getC_OrderLine_ID();
+				.filter(receiptLineTofind -> {
+					return receiptLineTofind.getC_OrderLine_ID() == purchaseOrderLine.getC_OrderLine_ID();
 				})
 				.findFirst();
 
 
-			MInOutLine shipmentLine = null;
-			if (maybeShipmentLine.isPresent()) {
-				shipmentLine = maybeShipmentLine.get();
-				BigDecimal orderQuantity = shipmentLine.getQtyEntered();
+			MInOutLine receiptLine = null;
+			if (maybeReceiptLine.isPresent()) {
+				receiptLine = maybeReceiptLine.get();
+				BigDecimal orderQuantity = receiptLine.getQtyEntered();
 				orderQuantity = orderQuantity.add(quantity);
-				shipmentLine.setQty(orderQuantity);
+				receiptLine.setQty(orderQuantity);
 			} else {
-				shipmentLine = new MInOutLine(shipment);
-				shipmentLine.setOrderLine(salesOrderLine, 0, quantity);
-				shipmentLine.setQty(quantity);
+				receiptLine = new MInOutLine(receipt);
+				receiptLine.setOrderLine(purchaseOrderLine, 0, quantity);
+				receiptLine.setQty(quantity);
 			}
-			shipmentLine.saveEx(transactionName);
+			receiptLine.saveEx(transactionName);
 
-			shipmentLineReference.set(shipmentLine);
+			receiptLineReference.set(receiptLine);
 		});
 
-		ShipmentLine.Builder builder = convertShipmentLine(
-			shipmentLineReference.get()
+		ReceiptLine.Builder builder = convertReceiptLine(
+			receiptLineReference.get()
 		);
 
 		return builder;
@@ -752,17 +766,18 @@ public class ExpressShipment extends ExpressShipmentImplBase {
 
 
 	@Override
-	public void deleteShipmentLine(DeleteShipmentLineRequest request, StreamObserver<Empty> responseObserver) {
+	public void deleteReceiptLine(DeleteReceiptLineRequest request, StreamObserver<Empty> responseObserver) {
 		try {
 			if (request == null) {
 				throw new AdempiereException("Object Request Null");
 			}
 
-			Empty.Builder builder = deleteShipmentLine(request);
+			Empty.Builder builder = deleteReceiptLine(request);
 			responseObserver.onNext(builder.build());
 			responseObserver.onCompleted();
 		} catch (Exception e) {
 			log.severe(e.getLocalizedMessage());
+			e.printStackTrace();
 			responseObserver.onError(Status.INTERNAL
 				.withDescription(e.getLocalizedMessage())
 				.withCause(e)
@@ -771,25 +786,22 @@ public class ExpressShipment extends ExpressShipmentImplBase {
 		}
 	}
 
-	private Empty.Builder deleteShipmentLine(DeleteShipmentLineRequest request) {
-		if (request.getId() <= 0) {
-			throw new AdempiereException("@M_InOutLine_ID@ @NotFound@");
+	private Empty.Builder deleteReceiptLine(DeleteReceiptLineRequest request) {
+		int receiptLineId = request.getId();
+		if (receiptLineId <= 0) {
+			throw new AdempiereException("@FillMandatory@ @M_InOutLine_ID@");
 		}
 
 		Trx.run(transactionName -> {
-			int shipmentLineId = request.getId();
-			if (shipmentLineId <= 0) {
-				throw new AdempiereException("@M_InOutLine_ID@ @NotFound@");
-			}
-			MInOutLine shipmentLine = new MInOutLine(Env.getCtx(), shipmentLineId, transactionName);
-			if (shipmentLine == null || shipmentLine.getM_InOutLine_ID() <= 0) {
+			MInOutLine receiptLine = new MInOutLine(Env.getCtx(), receiptLineId, transactionName);
+			if (receiptLine == null || receiptLine.getM_InOutLine_ID() <= 0) {
 				throw new AdempiereException("@M_InOutLine_ID@ @NotFound@");
 			}
 			// Validate processed
-			if (shipmentLine.isProcessed()) {
+			if (receiptLine.isProcessed()) {
 				throw new AdempiereException("@M_InOutLine_ID@ @Processed@");
 			}
-			shipmentLine.deleteEx(true);
+			receiptLine.deleteEx(true);
 		});
 
 		return Empty.newBuilder();
@@ -798,17 +810,18 @@ public class ExpressShipment extends ExpressShipmentImplBase {
 
 
 	@Override
-	public void updateShipmentLine(UpdateShipmentLineRequest request, StreamObserver<ShipmentLine> responseObserver) {
+	public void updateReceiptLine(UpdateReceiptLineRequest request, StreamObserver<ReceiptLine> responseObserver) {
 		try {
 			if (request == null) {
 				throw new AdempiereException("Object Request Null");
 			}
 
-			ShipmentLine.Builder builder = updateShipmentLine(request);
+			ReceiptLine.Builder builder = updateReceiptLine(request);
 			responseObserver.onNext(builder.build());
 			responseObserver.onCompleted();
 		} catch (Exception e) {
 			log.severe(e.getLocalizedMessage());
+			e.printStackTrace();
 			responseObserver.onError(Status.INTERNAL
 				.withDescription(e.getLocalizedMessage())
 				.withCause(e)
@@ -817,29 +830,26 @@ public class ExpressShipment extends ExpressShipmentImplBase {
 		}
 	}
 
-	private ShipmentLine.Builder updateShipmentLine(UpdateShipmentLineRequest request) {
-		if (request.getId() <= 0) {
-			throw new AdempiereException("@M_InOutLine_ID@ @NotFound@");
+	private ReceiptLine.Builder updateReceiptLine(UpdateReceiptLineRequest request) {
+		int receiptLineId = request.getId();
+		if (receiptLineId <= 0) {
+			throw new AdempiereException("@FillMandatory@ @M_InOutLine_ID@");
 		}
 
-		AtomicReference<MInOutLine> maybeShipmentLine = new AtomicReference<MInOutLine>();
+		AtomicReference<MInOutLine> maybeReceiptLine = new AtomicReference<MInOutLine>();
 
 		Trx.run(transactionName -> {
-			int shipmentLineId = request.getId();
-			if (shipmentLineId <= 0) {
-				throw new AdempiereException("@M_InOutLine_ID@ @NotFound@");
-			}
-			MInOutLine shipmentLine = new MInOutLine(Env.getCtx(), shipmentLineId, transactionName);
-			if (shipmentLine == null || shipmentLine.getM_InOutLine_ID() <= 0) {
+			MInOutLine receiptLine = new MInOutLine(Env.getCtx(), receiptLineId, transactionName);
+			if (receiptLine == null || receiptLine.getM_InOutLine_ID() <= 0) {
 				throw new AdempiereException("@M_InOutLine_ID@ @NotFound@");
 			}
 			// Validate processed
-			if (shipmentLine.isProcessed()) {
+			if (receiptLine.isProcessed()) {
 				throw new AdempiereException("@M_InOutLine_ID@ @Processed@");
 			}
 
-			MOrderLine salesOrderLine = new MOrderLine(Env.getCtx(), shipmentLine.getC_OrderLine_ID(), transactionName);
-			if (salesOrderLine == null || salesOrderLine.getC_OrderLine_ID() <= 0) {
+			MOrderLine purchaseOrderLine = new MOrderLine(Env.getCtx(), receiptLine.getC_OrderLine_ID(), transactionName);
+			if (purchaseOrderLine == null || purchaseOrderLine.getC_OrderLine_ID() <= 0) {
 				throw new AdempiereException("@C_OrderLine_ID@ @NotFound@");
 			}
 
@@ -852,23 +862,31 @@ public class ExpressShipment extends ExpressShipmentImplBase {
 					quantity = BigDecimal.ONE;
 				}
 			}
+
+			boolean validateOrderedQty = MSysConfig.getBooleanValue(
+				"VALIDATE_MATCHING_TO_ORDERED_QTY",
+				true,
+				Env.getAD_Client_ID(Env.getCtx())
+			);
 			// Validate available
-			BigDecimal orderQuantityDelivered = salesOrderLine.getQtyOrdered().subtract(salesOrderLine.getQtyDelivered());
-			if (orderQuantityDelivered.compareTo(quantity) < 0) {
-				throw new AdempiereException("@QtyInsufficient@");
+			if (validateOrderedQty) {
+				BigDecimal orderQuantityDelivered = purchaseOrderLine.getQtyOrdered().subtract(purchaseOrderLine.getQtyDelivered());
+				if (orderQuantityDelivered.compareTo(quantity) < 0) {
+					throw new AdempiereException("@QtyInsufficient@");
+				}
 			}
 
-			shipmentLine.setQty(quantity);
-			shipmentLine.setDescription(
+			receiptLine.setQty(quantity);
+			receiptLine.setDescription(
 				ValueManager.validateNull(request.getDescription())
 			);
-			shipmentLine.saveEx(transactionName);
+			receiptLine.saveEx(transactionName);
 
-			maybeShipmentLine.set(shipmentLine);
+			maybeReceiptLine.set(receiptLine);
 		});
 
-		ShipmentLine.Builder builder = convertShipmentLine(
-			maybeShipmentLine.get()
+		ReceiptLine.Builder builder = convertReceiptLine(
+			maybeReceiptLine.get()
 		);
 
 		return builder;
@@ -877,17 +895,18 @@ public class ExpressShipment extends ExpressShipmentImplBase {
 
 
 	@Override
-	public void listShipmentLines(ListShipmentLinesRequest request, StreamObserver<ListShipmentLinesResponse> responseObserver) {
+	public void listReceiptLines(ListReceiptLinesRequest request, StreamObserver<ListReceiptLinesResponse> responseObserver) {
 		try {
 			if (request == null) {
 				throw new AdempiereException("Object Request Null");
 			}
 
-			ListShipmentLinesResponse.Builder builderList = listShipmentLines(request);
+			ListReceiptLinesResponse.Builder builderList = listReceiptLines(request);
 			responseObserver.onNext(builderList.build());
 			responseObserver.onCompleted();
 		} catch (Exception e) {
 			log.severe(e.getLocalizedMessage());
+			e.printStackTrace();
 			responseObserver.onError(Status.INTERNAL
 				.withDescription(e.getLocalizedMessage())
 				.withCause(e)
@@ -896,27 +915,27 @@ public class ExpressShipment extends ExpressShipmentImplBase {
 		}
 	}
 
-	ShipmentLine.Builder convertShipmentLine(MInOutLine shipmentLine) {
-		ShipmentLine.Builder builder = ShipmentLine.newBuilder();
-		if (shipmentLine == null || shipmentLine.getM_InOutLine_ID() <= 0) {
+	ReceiptLine.Builder convertReceiptLine(MInOutLine receiptLine) {
+		ReceiptLine.Builder builder = ReceiptLine.newBuilder();
+		if (receiptLine == null || receiptLine.getM_InOutLine_ID() <= 0) {
 			return builder;
 		}
 
-		builder.setId(shipmentLine.getM_InOutLine_ID())
+		builder.setId(receiptLine.getM_InOutLine_ID())
 			.setProduct(
-				convertProduct(shipmentLine.getM_Product_ID())
+				convertProduct(receiptLine.getM_Product_ID())
 			)
 			.setDescription(
-				ValueManager.validateNull(shipmentLine.getDescription())
+				ValueManager.validateNull(receiptLine.getDescription())
 			)
 			.setQuantity(
 				NumberManager.getBigDecimalToString(
-					shipmentLine.getQtyEntered()
+					receiptLine.getQtyEntered()
 				)
 			)
-			.setLine(shipmentLine.getLine())
+			.setLine(receiptLine.getLine())
 		;
-		MOrderLine orderLine = new MOrderLine(Env.getCtx(), shipmentLine.getC_OrderLine_ID(), null);
+		MOrderLine orderLine = new MOrderLine(Env.getCtx(), receiptLine.getC_OrderLine_ID(), null);
 		if (orderLine != null && orderLine.getC_OrderLine_ID() > 0) {
 			builder.setOrderLineId(orderLine.getC_OrderLine_ID())
 			;
@@ -925,24 +944,26 @@ public class ExpressShipment extends ExpressShipmentImplBase {
 		return builder;
 	}
 
-	private ListShipmentLinesResponse.Builder listShipmentLines(ListShipmentLinesRequest request) {
-		if (request.getShipmentId() <= 0) {
+	private ListReceiptLinesResponse.Builder listReceiptLines(ListReceiptLinesRequest request) {
+		int receiptId = request.getReceiptId();
+		if (receiptId <= 0) {
+			throw new AdempiereException("@FillMandatory@ @M_InOut_ID@");
+		}
+		MInOut receipt = new MInOut(Env.getCtx(), receiptId, null);
+		if (receipt == null || receipt.getM_InOut_ID() <= 0) {
 			throw new AdempiereException("@M_InOut_ID@ @NotFound@");
 		}
-		int shipmentId = request.getShipmentId();
-		if (shipmentId <= 0) {
-			throw new AdempiereException("@M_InOut_ID@ @NotFound@");
-		}
+
 		Query query = new Query(
 			Env.getCtx(),
 			I_M_InOutLine.Table_Name,
 			I_M_InOutLine.COLUMNNAME_M_InOut_ID + " = ?",
 			null
 		)
-			.setParameters(shipmentId)
+			.setParameters(receiptId)
 			.setClient_ID()
-			.setOnlyActiveRecords(true)
 			.setApplyAccessFilter(MRole.SQL_FULLYQUALIFIED, MRole.SQL_RO)
+			.setOnlyActiveRecords(true)
 		;
 
 		int count = query.count();
@@ -955,7 +976,7 @@ public class ExpressShipment extends ExpressShipmentImplBase {
 			nexPageToken = LimitUtil.getPagePrefix(SessionManager.getSessionUuid()) + (pageNumber + 1);
 		}
 
-		ListShipmentLinesResponse.Builder builderList = ListShipmentLinesResponse.newBuilder()
+		ListReceiptLinesResponse.Builder builderList = ListReceiptLinesResponse.newBuilder()
 			.setRecordCount(count)
 			.setNextPageToken(
 				ValueManager.validateNull(nexPageToken)
@@ -963,9 +984,10 @@ public class ExpressShipment extends ExpressShipmentImplBase {
 		;
 
 		query.setLimit(limit, offset)
-			.list(MInOutLine.class)
-			.forEach(shipmentLine -> {
-				ShipmentLine.Builder builder = convertShipmentLine(shipmentLine);
+			.getIDsAsList()
+			.forEach(receiptLineId -> {
+				MInOutLine receiptLine = MInOutLine.get(Env.getCtx(), receiptLineId);
+				ReceiptLine.Builder builder = convertReceiptLine(receiptLine);
 				builderList.addRecords(builder);
 			});
 		;
