+++ src/main/java/org/spin/server/AllInOneServices.java
@@ -32,9 +32,6 @@ import org.spin.grpc.service.CoreFunctionality;
 import org.spin.grpc.service.Dashboarding;
 import org.spin.grpc.service.Dictionary;
 import org.spin.grpc.service.Enrollment;
-import org.spin.grpc.service.ExpressMovement;
-import org.spin.grpc.service.ExpressReceipt;
-import org.spin.grpc.service.ExpressShipment;
 import org.spin.grpc.service.FileManagement;
 import org.spin.grpc.service.GeneralLedger;
 import org.spin.grpc.service.ImportFileLoader;
@@ -45,7 +42,6 @@ import org.spin.grpc.service.LogsInfo;
 import org.spin.grpc.service.MatchPOReceiptInvoice;
 import org.spin.grpc.service.MaterialManagement;
 import org.spin.grpc.service.OrderInfo;
-import org.spin.grpc.service.PaymentAllocation;
 import org.spin.grpc.service.PaymentPrintExport;
 import org.spin.grpc.service.PaymentInfo;
 import org.spin.grpc.service.PayrollActionNotice;
@@ -62,6 +58,10 @@ import org.spin.grpc.service.UserInterface;
 import org.spin.grpc.service.WebStore;
 import org.spin.grpc.service.Workflow;
 import org.spin.grpc.service.field.location_address.LocationAddress;
+import org.spin.grpc.service.form.ExpressMovement;
+import org.spin.grpc.service.form.ExpressReceipt;
+import org.spin.grpc.service.form.ExpressShipment;
+import org.spin.grpc.service.form.PaymentAllocation;
 import org.spin.service.grpc.authentication.AuthorizationServerInterceptor;
 import org.spin.service.grpc.context.ServiceContextProvider;
